refactor: move webhook cases to modules
